Split database testing into its own package
In order to be able to do postgres database cleanup from tests, it now
needs to be part of the database interface.

Note that the database is unusable after `Drop()`ing it
